# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sun Proactive AI — Autonomous Social Impact Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /auth/onboarding:
    post:
      operationId: submitOnboarding
      tags: [auth]
      summary: Submit volunteer onboarding survey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingRequest"
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # Tasks
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # AI Endpoints
...
  /openai/conversations:
    get:
      operationId: getConversations
      tags: [openai]
      summary: Get user conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [openai]
      summary: Create a new conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /openai/conversations/{id}:
    get:
      operationId: getConversation
      tags: [openai]
      summary: Get conversation with messages
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationWithMessages"
  /openai/conversations/{id}/messages:
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a message (SSE streaming)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [volunteer, admin]
        language:
          type: string
          enum: [kz, ru, en]
          default: ru
      required: [email, password, name]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]
    OnboardingRequest:
      type: object
      properties:
        interests:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        availability:
          type: string
        location:
          type: string
        ageGroup:
          type: string
        transportMode:
          type: string
        answers:
          type: object
          additionalProperties: true
      required: [interests, skills]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [volunteer, admin]
        language:
          type: string
          enum: [kz, ru, en]
        avatar:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        level:
          type: integer
        xp:
          type: integer
        badges:
          type: array
          items:
            type: string
        totalHours:
          type: number
        rating:
          type: number
        skills:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        location:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, language, onboardingCompleted, level, xp, badges, totalHours, rating, skills, interests, createdAt]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        taskId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, message, read, createdAt]
...
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        userId:
          type: integer
        taskId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, userId, createdAt, updatedAt]
    ConversationWithMessages:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
          required: [messages]
    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, role, content, createdAt]
    CreateConversationRequest:
      type: object
      properties:
        title:
          type: string
        taskId:
          type: integer
          nullable: true
      required: [title]
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
        language:
          type: string
          enum: [kz, ru, en]
          default: ru
        systemContext:
          type: string
          nullable: true
      required: [content]